Move cairo_run method
