+++ src/vm_core.rs
@@ -1,27 +1,37 @@
 use crate::ids::PyIds;
 use crate::pycell;
 use crate::scope_manager::{PyEnterScope, PyExitScope};
+use crate::utils::to_py_error;
 use crate::{
     memory::PyMemory, memory_segments::PySegmentManager, relocatable::PyRelocatable,
     utils::to_vm_error,
 };
 use cairo_rs::any_box;
+use cairo_rs::cairo_run::write_output;
+use cairo_rs::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::BuiltinHintProcessor;
 use cairo_rs::hint_processor::hint_processor_definition::HintProcessor;
 use cairo_rs::hint_processor::proxies::exec_scopes_proxy::{
     get_exec_scopes_proxy, ExecutionScopesProxy,
 };
 use cairo_rs::types::exec_scope::ExecutionScopes;
+use cairo_rs::types::program::Program;
+use cairo_rs::types::relocatable::Relocatable;
+use cairo_rs::vm::errors::cairo_run_errors::CairoRunError;
+use cairo_rs::vm::errors::runner_errors::RunnerError;
+use cairo_rs::vm::errors::trace_errors::TraceError;
+use cairo_rs::vm::runners::cairo_runner::CairoRunner;
 use cairo_rs::vm::vm_core::VirtualMachine;
 use cairo_rs::{
     hint_processor::builtin_hint_processor::builtin_hint_processor_definition::HintProcessorData,
     vm::errors::vm_errors::VirtualMachineError,
 };
 use num_bigint::BigInt;
-use pyo3::PyCell;
 use pyo3::{pyclass, pymethods, PyObject, ToPyObject};
 use pyo3::{types::PyDict, Python};
+use pyo3::{PyCell, PyResult};
 use std::any::Any;
 use std::collections::HashMap;
+use std::path::{Path, PathBuf};
 use std::{cell::RefCell, rc::Rc};
 
 #[pyclass(unsendable)]
@@ -36,6 +46,64 @@ impl PyVM {
         PyVM {
             vm: Rc::new(RefCell::new(VirtualMachine::new(prime, trace_enabled))),
         }
+    }
+
+    #[pyo3(name = "cairo_run")]
+    pub fn cairo_run_py<'a>(
+        &self,
+        path: &'a str,
+        entrypoint: &'a str,
+        print_output: bool,
+        trace_file: Option<&str>,
+        memory_file: Option<&str>,
+    ) -> PyResult<()> {
+        let path = Path::new(path);
+        let program = Program::new(path, entrypoint).map_err(to_py_error)?;
+        let hint_processor = BuiltinHintProcessor::new_empty();
+        let mut cairo_runner = CairoRunner::new(&program, &hint_processor).map_err(to_py_error)?;
+        let end = cairo_runner
+            .initialize(&mut self.vm.borrow_mut())
+            .map_err(to_py_error)?;
+
+        self.run_until_pc(&mut cairo_runner, end)
+            .map_err(to_py_error)?;
+
+        self.vm
+            .borrow_mut()
+            .verify_auto_deductions()
+            .map_err(to_py_error)?;
+
+        cairo_runner
+            .relocate(&mut self.vm.borrow_mut())
+            .map_err(to_py_error)?;
+
+        if print_output {
+            write_output(&mut cairo_runner, &mut self.vm.borrow_mut()).map_err(to_py_error)?;
+        }
+
+        if let Some(trace_path) = trace_file {
+            let trace_path = PathBuf::from(trace_path);
+            let relocated_trace = cairo_runner
+                .relocated_trace
+                .as_ref()
+                .ok_or(CairoRunError::Trace(TraceError::TraceNotEnabled))
+                .map_err(to_py_error)?;
+
+            match cairo_rs::cairo_run::write_binary_trace(relocated_trace, &trace_path) {
+                Ok(()) => (),
+                Err(_e) => {
+                    return Err(CairoRunError::Runner(RunnerError::WriteFail)).map_err(to_py_error)
+                }
+            }
+        }
+
+        if let Some(memory_path) = memory_file {
+            let memory_path = PathBuf::from(memory_path);
+            cairo_rs::cairo_run::write_binary_memory(&cairo_runner.relocated_memory, &memory_path)
+                .map_err(|_| to_py_error(CairoRunError::Runner(RunnerError::WriteFail)))?;
+        }
+
+        Ok(())
     }
 }
 
@@ -145,6 +213,24 @@ impl PyVM {
             Err(VirtualMachineError::UnknownHint(_)) => Ok(true),
             Err(e) => Err(e),
         }
+    }
+
+    fn run_until_pc(
+        &self,
+        cairo_runner: &mut CairoRunner,
+        address: Relocatable,
+    ) -> Result<(), VirtualMachineError> {
+        let references = cairo_runner.get_reference_list();
+        let hint_data_dictionary = cairo_runner.get_hint_data_dictionary(&references)?;
+
+        while self.vm.borrow().get_pc() != &address {
+            self.step(
+                cairo_runner.hint_executor,
+                &mut cairo_runner.exec_scopes,
+                &hint_data_dictionary,
+            )?;
+        }
+        Ok(())
     }
 }
